@@ -1,45 +1,61 @@
-import React, { Component } from 'react';
+import React, { useState, useEffect } from 'react';
 import CardList from '../components/CardList';
 import SearchBox from '../components/SearchBox';
 import Scroll from '../components/Scroll';
 import './App.css';
+// hooks topic
+// class App extends Component
+function App() {
+  // constructor() {
+  //   super()
+  //   this.state = {
+  //     robots: [],
+  //     searchfield: ''
+  //   }
+  // }
+  //
+  // componentDidMount() {
+  //   fetch('https://jsonplaceholder.typicode.com/users')
+  //     .then(response => response.json())
+  //     .then(users => {this.setState({ robots: users})});
+  // }
 
-class App extends Component {
-  constructor() {
-    super()
-    this.state = {
-      robots: [],
-      searchfield: ''
-    }
+  const [robots, setRobots] = useState([]);
+  const [searchfield, setSearchfield] = useState('');
+  const [count, setCount] = useState(0);
+
+  useEffect(() =>{
+      fetch('https://jsonplaceholder.typicode.com/users')
+        .then(response => response.json())
+        .then(users => setRobots(users));
+    // console.log(robots, searchfield);
+    console.log(count);
+  }, [count]) // [count] with count useEffect runs when count chaanges
+
+  const onSearchChange = (event) => {
+    // this.setState({ searchfield: event.target.value })
+    setSearchfield(event.target.value);
   }
 
-  componentDidMount() {
-    fetch('https://jsonplaceholder.typicode.com/users')
-      .then(response=> response.json())
-      .then(users => {this.setState({ robots: users})});
-  }
 
-  onSearchChange = (event) => {
-    this.setState({ searchfield: event.target.value })
-  }
-
-  render() {
-    const { robots, searchfield } = this.state;
+    // const { robots, searchfield } = this.state;
     const filteredRobots = robots.filter(robot =>{
       return robot.name.toLowerCase().includes(searchfield.toLowerCase());
     })
+
     return !robots.length ?
       <h1>Loading</h1> :
       (
         <div className='tc'>
           <h1 className='f1'>RoboFriends</h1>
-          <SearchBox searchChange={this.onSearchChange}/>
+          <button onClick={()=>setCount(count+1)}>Click!</button>
+          <SearchBox searchChange={onSearchChange}/>
           <Scroll>
             <CardList robots={filteredRobots} />
           </Scroll>
         </div>
       );
-  }
+
 }
 
 export default App;
